Fastapi_update/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
redis==5.2.1
python-multipart==0.0.20
websockets==15.0.1
requests==2.32.3
PyJWT==2.9.0
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
prometheus-client==0.21.1
boto3==1.35.71
python-dotenv==1.0.1
